requirements.txt
streamlit
google-generativeai
python-docx
docx
ipykernel
numpy
matplotlib
pandas
